requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_tflow_select=2.3.0=mkl
absl-py=0.9.0=py38_0
asgiref=3.2.7=py_0
astunparse=1.6.3=py_0
blas=1.0=mkl
blinker=1.4=py38_0
brotlipy=0.7.0=py38h7b6447c_1000
bzip2=1.0.8=h516909a_2
c-ares=1.15.0=h7b6447c_1001
ca-certificates=2020.6.20=hecda079_0
cachetools=4.1.0=py_1
cairo=1.16.0=hcf35c78_1003
certifi=2020.6.20=py38h32f6830_0
cffi=1.14.0=py38hd463f26_0
chardet=3.0.4=py38_1003
click=7.1.2=py_0
cryptography=2.9.2=py38h1ba5d50_0
dbus=1.13.6=he372182_0
django=3.0.3=py_0
djangorestframework=3.11.0=pypi_0
expat=2.2.9=he1b5a44_2
ffmpeg=4.2.3=h167e202_0
fontconfig=2.13.1=h86ecdb6_1001
freetype=2.10.2=he06d7ca_0
gast=0.3.3=py_0
gettext=0.19.8.1=hc5be6a0_1002
glib=2.65.0=h6f030ca_0
gmp=6.2.0=he1b5a44_2
gnutls=3.6.13=h79a8f9a_0
google-auth=1.14.1=py_0
google-auth-oauthlib=0.4.1=py_2
google-pasta=0.2.0=py_0
graphite2=1.3.13=he1b5a44_1001
grpcio=1.27.2=py38hf8bcb03_0
gst-plugins-base=1.14.5=h0935bb2_2
gstreamer=1.14.5=h36ae1b5_2
h5py=2.10.0=py38hd6299e0_1
harfbuzz=2.4.0=h9f30f68_3
hdf5=1.10.6=hb1b8bf9_0
icu=64.2=he1b5a44_1
idna=2.10=py_0
intel-openmp=2020.1=217
jasper=1.900.1=h07fcdf6_1006
jpeg=9d=h516909a_0
keras=2.4.3=0
keras-base=2.4.3=py_0
keras-preprocessing=1.1.0=py_1
lame=3.100=h14c3975_1001
ld_impl_linux-64=2.33.1=h53a641e_7
libblas=3.8.0=16_mkl
libcblas=3.8.0=16_mkl
libclang=9.0.1=default_hde54327_0
libedit=3.1.20191231=h7b6447c_0
libffi=3.2.1=he1b5a44_1007
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libiconv=1.15=h516909a_1006
liblapack=3.8.0=16_mkl
liblapacke=3.8.0=16_mkl
libllvm9=9.0.1=he513fc3_1
libopencv=4.3.0=py38_1
libpng=1.6.37=hed695b0_1
libprotobuf=3.12.3=hd408876_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=hc7e4089_6
libuuid=2.32.1=h14c3975_1000
libwebp-base=1.1.0=h516909a_3
libxcb=1.13=h14c3975_1002
libxkbcommon=0.10.0=he1b5a44_0
libxml2=2.9.10=hee79883_0
lz4-c=1.9.2=he1b5a44_1
markdown=3.1.1=py38_0
mkl=2020.1=217
mkl-service=2.3.0=py38he904b0f_0
mkl_fft=1.1.0=py38h23d657b_0
mkl_random=1.1.1=py38h0573a6f_0
ncurses=6.2=he6710b0_1
nettle=3.4.1=h1bed415_1002
nspr=4.26=he1b5a44_0
nss=3.47=he751ad9_0
numpy=1.18.5=py38ha1c710e_0
numpy-base=1.18.5=py38hde5b4d6_0
oauthlib=3.1.0=py_0
opencv=4.3.0=py38_1
openh264=2.1.1=h8b12597_0
openssl=1.1.1g=h516909a_0
opt_einsum=3.1.0=py_0
pcre=8.44=he1b5a44_0
pip=20.1.1=py38_1
pixman=0.38.0=h516909a_1003
protobuf=3.12.3=py38he6710b0_0
pthread-stubs=0.4=h14c3975_1001
py-opencv=4.3.0=py38h23f93f0_1
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.7=py_0
pycparser=2.20=py_0
pyjwt=1.7.1=py38_0
pyopenssl=19.1.0=py38_0
pysocks=1.7.1=py38_0
python=3.8.3=cpython_he5300dc_0
python_abi=3.8=1_cp38
pytz=2020.1=py_0
pyyaml=5.3.1=py38h7b6447c_1
qt=5.12.5=hd8c4c69_1
readline=8.0=h7b6447c_0
requests=2.24.0=py_0
requests-oauthlib=1.3.0=py_0
rsa=4.0=py_0
scipy=1.5.0=py38h0b6359f_0
semantics3=0.3.9=pypi_0
setuptools=47.3.1=py38_0
six=1.15.0=py_0
sqlite=3.32.3=h62c20be_0
sqlparse=0.3.1=py_0
tensorboard=2.2.1=pyh532a8cf_0
tensorboard-plugin-wit=1.6.0=py_0
tensorflow=2.2.0=mkl_py38h6d3daf0_0
tensorflow-base=2.2.0=mkl_py38h5059a2d_0
tensorflow-estimator=2.2.0=pyh208ff02_0
termcolor=1.1.0=py38_1
tk=8.6.10=hbc83047_0
url-normalize=1.4.2=pypi_0
urllib3=1.25.9=py_0
werkzeug=1.0.1=py_0
wheel=0.34.2=py38_0
wrapt=1.12.1=py38h7b6447c_1
x264=1!152.20180806=h14c3975_0
xorg-kbproto=1.0.7=h14c3975_1002
xorg-libice=1.0.10=h516909a_0
xorg-libsm=1.2.3=h84519dc_1000
xorg-libx11=1.6.9=h516909a_0
xorg-libxau=1.0.9=h14c3975_0
xorg-libxdmcp=1.1.3=h516909a_0
xorg-libxext=1.3.4=h516909a_0
xorg-libxrender=0.9.10=h516909a_1002
xorg-renderproto=0.11.1=h14c3975_1002
xorg-xextproto=7.3.0=h14c3975_1002
xorg-xproto=7.0.31=h14c3975_1007
xz=5.2.5=h7b6447c_0
yaml=0.2.5=h7b6447c_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.4=h6597ccf_3
